# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ExportPreviewResponse": {
        "properties": {
          "blocking_errors": {
            "items": {
              "type": "string"
            },
            "title": "Blocking Errors",
            "type": "array"
          },
          "dataset_version_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Dataset Version Ids",
            "type": "array"
          },
          "distance_m": {
            "title": "Distance M",
            "type": "number"
          },
          "edge_count": {
            "title": "Edge Count",
            "type": "integer"
          },
          "journey_count": {
            "title": "Journey Count",
            "type": "integer"
          },
          "preview_token": {
            "title": "Preview Token",
            "type": "string"
          },
          "rail_dataset_version_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Rail Dataset Version Ids",
            "type": "array"
          },
          "rail_graph_versions": {
            "items": {
              "type": "string"
            },
            "title": "Rail Graph Versions",
            "type": "array"
          },
          "segment_count": {
            "title": "Segment Count",
            "type": "integer"
          },
          "track_count": {
            "title": "Track Count",
            "type": "integer"
          },
          "unique_edge_count": {
            "title": "Unique Edge Count",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "preview_token",
          "journey_count",
          "edge_count",
          "unique_edge_count",
          "distance_m",
          "track_count",
          "segment_count",
          "dataset_version_ids",
          "rail_dataset_version_ids",
          "rail_graph_versions",
          "blocking_errors",
          "warnings"
        ],
        "title": "ExportPreviewResponse",
        "type": "object"
      },
      "ExportRequest": {
        "properties": {
          "city_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "journey_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 5000,
            "title": "Journey Ids",
            "type": "array"
          },
          "line_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Id"
          },
          "max_segment_length_m": {
            "anyOf": [
              {
                "enum": [
                  15,
                  25,
                  50
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 25,
            "title": "Max Segment Length M"
          },
          "mode": {
            "default": "coverage",
            "enum": [
              "journeys",
              "coverage"
            ],
            "title": "Mode",
            "type": "string"
          },
          "rail_max_segment_length_m": {
            "anyOf": [
              {
                "enum": [
                  100,
                  200,
                  500
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 200,
            "title": "Rail Max Segment Length M"
          },
          "traveled_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traveled From"
          },
          "traveled_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traveled To"
          }
        },
        "title": "ExportRequest",
        "type": "object"
      },
...
      "GPXRequest": {
        "properties": {
          "city_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "journey_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 5000,
            "title": "Journey Ids",
            "type": "array"
          },
          "line_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Id"
          },
          "max_segment_length_m": {
            "anyOf": [
              {
                "enum": [
                  15,
                  25,
                  50
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 25,
            "title": "Max Segment Length M"
          },
          "mode": {
            "default": "coverage",
            "enum": [
              "journeys",
              "coverage"
            ],
            "title": "Mode",
            "type": "string"
          },
          "preview_token": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Preview Token",
            "type": "string"
          },
          "rail_max_segment_length_m": {
            "anyOf": [
              {
                "enum": [
                  100,
                  200,
                  500
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 200,
            "title": "Rail Max Segment Length M"
          },
          "traveled_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traveled From"
          },
          "traveled_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traveled To"
          }
        },
        "required": [
          "preview_token"
        ],
        "title": "GPXRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache License 2.0"
    },
    "title": "Transit2GPX API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/exports/gpx": {
      "post": {
        "operationId": "download_gpx_api_v1_exports_gpx_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GPXRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Gpx",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/preview": {
      "post": {
        "operationId": "preview_export_api_v1_exports_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Export",
        "tags": [
          "exports"
        ]
      }
    },
...
  }
}